+++ lib/widgets/list_items/melody_item.dart
@@ -12,9 +12,10 @@ import 'package:flutter/material.dart';
 class MelodyItem extends StatefulWidget {
   final Melody? melody;
   final BuildContext? context;
-  final bool? isRounded;
-  final double? imageSize;
-  final double? padding;
+  final bool isRounded;
+  final double imageSize;
+  final double padding;
+
   MelodyItem(
       {Key? key,
       this.melody,
@@ -38,7 +39,7 @@ class _MelodyItemState extends State<MelodyItem> {
 
   @override
   void initState() {
-    if (widget.melody!.authorId != null) {
+    if (widget.melody?.authorId != null) {
       getAuthor();
     }
     super.initState();
@@ -52,58 +53,64 @@ class _MelodyItemState extends State<MelodyItem> {
   }
 
   getAuthor() async {
-    User author = await DatabaseService.getUserWithId(widget.melody!.authorId!);
-    setState(() {
-      _author = author;
-    });
+    if (widget.melody?.authorId != null) {
+      User author =
+          await DatabaseService.getUserWithId(widget.melody!.authorId!);
+      setState(() {
+        _author = author;
+      });
+    }
   }
 
   isFavourite() async {
-    bool isFavourite = (await usersRef
-            .doc(Constants.currentUserID)
-            .collection('favourites')
-            .doc(widget.melody!.id)
-            .get())
-        .exists;
+    if (widget.melody != null && widget.melody!.id != null) {
+      bool isFavourite = (await usersRef
+              .doc(Constants.currentUserID)
+              .collection('favourites')
+              .doc(widget.melody!.id!)
+              .get())
+          .exists;
 
-    setState(() {
-      _isFavourite = isFavourite;
-    });
+      setState(() {
+        _isFavourite = isFavourite;
+      });
+    }
   }
 
   @override
   Widget build(BuildContext context) {
     return Padding(
-      padding: EdgeInsets.all(widget.padding!),
+      padding: EdgeInsets.all(widget.padding),
       child: Container(
         width: MediaQuery.of(context).size.width,
         decoration: BoxDecoration(
           color: Colors.white.withOpacity(.1),
-          borderRadius: widget.isRounded! ? BorderRadius.circular(20.0) : null,
+          borderRadius: widget.isRounded ? BorderRadius.circular(20.0) : null,
         ),
         child: ListTile(
           leading: CachedImage(
-            width: widget.imageSize!,
-            height: widget.imageSize!,
-            imageUrl: widget.melody!.imageUrl!,
+            width: widget.imageSize,
+            height: widget.imageSize,
+            imageUrl: widget.melody?.imageUrl,
             imageShape: BoxShape.rectangle,
             defaultAssetImage: Strings.default_melody_image,
           ),
-          title: Text(widget.melody!.name ?? '',
+          title: Text(widget.melody?.name ?? '',
               style: TextStyle(color: MyColors.textLightColor)),
           subtitle: Text(
-            _author?.name ?? widget.melody!.singer ?? '',
+            _author?.name ?? widget.melody?.singer ?? '',
             style: TextStyle(color: MyColors.textLightColor),
           ),
           trailing: widget.melody?.isSong ?? false
               ? InkWell(
                   onTap: () =>
                       AppUtil.executeFunctionIfLoggedIn(context, () async {
-                    _isFavourite
-                        ? await DatabaseService.deleteMelodyFromFavourites(
-                            widget.melody!.id!)
-                        : await DatabaseService.addMelodyToFavourites(
-                            widget.melody!.id!);
+                    if (widget.melody != null && widget.melody!.id != null)
+                      _isFavourite
+                          ? await DatabaseService.deleteMelodyFromFavourites(
+                              widget.melody!.id!)
+                          : await DatabaseService.addMelodyToFavourites(
+                              widget.melody!.id!);
 
                     await isFavourite();
                   }),
